Lot of refactoring and thread safe scopes stack
- MT_UNUSED for unused variables
- MT_USED_IN_ASSERT for variables used only in assertions
- MT_NOCOPYABLE instead of private assignment operators and private
ctros
- MT_ASSERT now is do nothing in release
- mt_thread_local for thread local variables
- thread safe scopes stack designed for use in debug tools (WiP)

+++ Scheduler/Include/MTTaskGroup.h
@@ -28,11 +28,11 @@ namespace MT
 	//Task group ID
 	class TaskGroup
 	{
-		int id;
+		int16 id;
 
 	public:
 
-		static const int MT_MAX_GROUPS_COUNT = 256;
+		static const int16 MT_MAX_GROUPS_COUNT = 256;
 
 		enum PredefinedValues
 		{
@@ -49,10 +49,10 @@ namespace MT
 
 		explicit TaskGroup(PredefinedValues v)
 		{
-			id = v;
+			id = (int16)v;
 		}
 
-		explicit TaskGroup(int _id)
+		explicit TaskGroup(int16 _id)
 		{
 			id = _id;
 		}
@@ -64,7 +64,7 @@ namespace MT
 
 		TaskGroup & operator= (const PredefinedValues & v)
 		{
-			id = v;
+			id = (int16)v;
 			return *this;
 		}
